// client/src/app/model/place.ts
// backend place
export interface Place {
    placeId?: number;      // Optional, generated by backend,  nothing entered = undefined
    tripId: number;
    placeName: string;
    address: string;
    latitude: number;
    longitude: number;
    dayNumber: number;
    orderIndex: number;
    createdAt?: Date;
}

// place on the map
export interface MapPlace 
{
    lat: number;
    lng: number;
    day?: number;   // maps to dayNumber in backend
    name?: string;  // maps to placeName in backend
    address?: string;
}

// place on the itinerary
export interface PlannedPlace {
    id: number;
    name?: string;
    address?: string;
    lat: number;
    lng: number;
    day: number;
    orderIndex: number;
  }